requirements.txt
requests>=2.31.0
markdown>=3.5
beautifulsoup4>=4.12.0
Pillow>=10.0.0
